+++ loan-module-service/src/main/java/com/example/loanmoduleservice/service/NotificationService.java
@@ -4,6 +4,8 @@ import com.example.loanmoduleservice.entity.RepaymentSchedule;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 @Slf4j
 public class NotificationService {
@@ -13,7 +15,7 @@ public class NotificationService {
         System.out.printf("Sending email to %s:\nSubject: %s\nMessage: %s\n", email, subject, message);
     }
 
-    public void notifyLoanApproval(String email, double loanAmount, double emi, int tenure) {
+    public void notifyLoanApproval(String email, BigDecimal loanAmount, BigDecimal emi, int tenure) {
         String subject = "Loan Approved";
         String message = String.format(
                 "Congratulations! Your loan of %.2f has been approved. Your EMI is %.2f for %d months.",
@@ -22,7 +24,7 @@ public class NotificationService {
         sendNotification(email, subject, message);
     }
 
-    public void notifyLoanRepaid(String email, double loanAmount, double emi, int tenure) {
+    public void notifyLoanRepaid(String email, BigDecimal loanAmount, BigDecimal emi, int tenure) {
         String subject = "Loan Approved";
         String message = String.format(
                 "Congratulations! Your loan of %.2f has been repaid. Your EMI is %.2f for %d months.",
@@ -36,7 +38,7 @@ public class NotificationService {
         sendNotification(email, subject, message);
     }
 
-    public void notifyLoanDisbursement(String email, Double amount, String accountNumber) {
+    public void notifyLoanDisbursement(String email, BigDecimal amount, String accountNumber) {
         String subject = "Loan Disbursed";
         String message = String.format("Your loan of %.2f has been disbursed to account %s.", amount, accountNumber);
         sendNotification(email, subject, message);
@@ -48,7 +50,7 @@ public class NotificationService {
                 schedule.getInstallmentAmount(), schedule.getDueDate());
     }
 
-    public void sendMissedPaymentNotificationWithPenalty(RepaymentSchedule schedule, Double penalty) {
+    public void sendMissedPaymentNotificationWithPenalty(RepaymentSchedule schedule, BigDecimal penalty) {
         System.out.printf("Reminder: Your payment of %.2f due on %s is overdue.%nPenalty incurred: %.2f.%n",
                 schedule.getInstallmentAmount(), schedule.getDueDate(), penalty);
     }
