+++ loan-module-service/src/main/java/com/example/loanmoduleservice/service/RepaymentService.java
@@ -5,11 +5,12 @@ import com.example.loanmoduleservice.entity.LoanApplication;
 import com.example.loanmoduleservice.entity.RepaymentSchedule;
 import com.fasterxml.jackson.core.JsonProcessingException;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface RepaymentService {
     public List<RepaymentSchedule> generateRepaymentSchedule(LoanApplication loanApplication);
-    public Double calculatePenalty(RepaymentSchedule schedule);
+    public BigDecimal calculatePenalty(RepaymentSchedule schedule);
 
     public RepaymentSchedule rePayment(Long repaymentScheduleId, Repayment repayment, Long userId) throws JsonProcessingException;
 }
