+++ loan-module-service/src/main/java/com/example/loanmoduleservice/service/TransactionLogService.java
@@ -5,10 +5,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 
 public interface TransactionLogService {
     public void recordTransaction(Long loanApplicationId, BankDetails payerBankDetails, BankDetails receiverBankDetails,
-                                  Double amount, String transactionId, String transactionType, Long userId) throws JsonProcessingException;
+                                  BigDecimal amount, String transactionId, String transactionType, Long userId) throws JsonProcessingException;
 
     public void exportToExcel(HttpServletResponse response, String userId, String startDate, String endDate) throws IOException;
 }
