+++ loan-module-service/src/main/java/com/example/loanmoduleservice/service/impl/FundTransferServiceImpl.java
@@ -4,10 +4,11 @@ import com.example.loanmoduleservice.dtos.request.BankDetails;
 import com.example.loanmoduleservice.service.FundTransferService;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 @Service
 public class FundTransferServiceImpl implements FundTransferService {
-    public String transferFunds(BankDetails senderBankDetails, BankDetails receiverBankDetails, Double amount) {
+    public String transferFunds(BankDetails senderBankDetails, BankDetails receiverBankDetails, BigDecimal amount) {
         // Simulate fund transfer via banking API
         System.out.printf("Transferring %.2f to account: %s\n", amount, receiverBankDetails.getAccountNumber());
 
@@ -15,7 +16,7 @@ public class FundTransferServiceImpl implements FundTransferService {
         return UUID.randomUUID().toString();
     }
 
-    public String transferFunds(BankDetails receiverBankDetails, Double amount){
+    public String transferFunds(BankDetails receiverBankDetails, BigDecimal amount){
         // Simulate fund transfer via banking API
         System.out.printf("Transferring %.2f to account: %s\n", amount, receiverBankDetails.getAccountNumber());
